+++ saleor/graphql/discount/tests/test_bulk_delete.py
@@ -1,3 +1,5 @@
+from unittest import mock
+
 import graphene
 import pytest
 
@@ -74,14 +76,49 @@ def test_delete_sales(staff_api_client, sale_list, permission_manage_discounts):
     assert not Sale.objects.filter(id__in=[sale.id for sale in sale_list]).exists()
 
 
-def test_delete_vouchers(staff_api_client, voucher_list, permission_manage_discounts):
-    query = """
+BULK_DELETE_VOUCHERS_MUTATION = """
     mutation voucherBulkDelete($ids: [ID!]!) {
         voucherBulkDelete(ids: $ids) {
             count
         }
     }
-    """
+"""
+
+
+def test_delete_vouchers(staff_api_client, voucher_list, permission_manage_discounts):
+    variables = {
+        "ids": [
+            graphene.Node.to_global_id("Voucher", voucher.id)
+            for voucher in voucher_list
+        ]
+    }
+    response = staff_api_client.post_graphql(
+        BULK_DELETE_VOUCHERS_MUTATION,
+        variables,
+        permissions=[permission_manage_discounts],
+    )
+    content = get_graphql_content(response)
+
+    assert content["data"]["voucherBulkDelete"]["count"] == 3
+    assert not Voucher.objects.filter(
+        id__in=[voucher.id for voucher in voucher_list]
+    ).exists()
+
+
+@mock.patch("saleor.plugins.webhook.plugin.get_webhooks_for_event")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
+def test_delete_vouchers_trigger_webhook(
+    mocked_webhook_trigger,
+    mocked_get_webhooks_for_event,
+    any_webhook,
+    staff_api_client,
+    voucher_list,
+    permission_manage_discounts,
+    settings,
+):
+    # given
+    mocked_get_webhooks_for_event.return_value = [any_webhook]
+    settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
 
     variables = {
         "ids": [
@@ -90,11 +127,11 @@ def test_delete_vouchers(staff_api_client, voucher_list, permission_manage_disco
         ]
     }
     response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_discounts]
+        BULK_DELETE_VOUCHERS_MUTATION,
+        variables,
+        permissions=[permission_manage_discounts],
     )
     content = get_graphql_content(response)
 
     assert content["data"]["voucherBulkDelete"]["count"] == 3
-    assert not Voucher.objects.filter(
-        id__in=[voucher.id for voucher in voucher_list]
-    ).exists()
+    assert mocked_webhook_trigger.call_count == len(voucher_list)