+++ saleor/plugins/webhook/plugin.py
@@ -30,6 +30,7 @@ from ...webhook.payloads import (
     generate_product_variant_with_stock_payload,
     generate_requestor,
     generate_sale_payload,
+    generate_transaction_action_request_payload,
     generate_translation_payload,
 )
 from ...webhook.utils import get_webhooks_for_event
@@ -52,13 +53,18 @@ if TYPE_CHECKING:
     from ...account.models import User
     from ...channel.models import Channel
     from ...checkout.models import Checkout
-    from ...discount.models import Sale
+    from ...discount.models import Sale, Voucher
     from ...giftcard.models import GiftCard
     from ...graphql.discount.mutations import NodeCatalogueInfo
     from ...invoice.models import Invoice
     from ...order.models import Fulfillment, Order
     from ...page.models import Page
-    from ...payment.interface import GatewayResponse, PaymentData, PaymentGateway
+    from ...payment.interface import (
+        GatewayResponse,
+        PaymentData,
+        PaymentGateway,
+        TransactionActionData,
+    )
     from ...product.models import Category, Collection, Product, ProductVariant
     from ...shipping.interface import ShippingMethodData
     from ...shipping.models import ShippingMethod, ShippingZone
@@ -744,11 +750,55 @@ class WebhookPlugin(BasePlugin):
                 translation_data, event_type, webhooks, translation, self.requestor
             )
 
+    def _trigger_voucher_event(self, event_type, voucher):
+        if webhooks := get_webhooks_for_event(event_type):
+            payload = {
+                "id": graphene.Node.to_global_id("Voucher", voucher.id),
+                "name": voucher.name,
+                "code": voucher.code,
+            }
+            trigger_webhooks_async(
+                payload, event_type, webhooks, voucher, self.requestor
+            )
+
+    def voucher_created(self, voucher: "Voucher", previous_value: None) -> None:
+        if not self.active:
+            return previous_value
+        self._trigger_voucher_event(WebhookEventAsyncType.VOUCHER_CREATED, voucher)
+
+    def voucher_updated(self, voucher: "Voucher", previous_value: None) -> None:
+        if not self.active:
+            return previous_value
+        self._trigger_voucher_event(WebhookEventAsyncType.VOUCHER_UPDATED, voucher)
+
+    def voucher_deleted(self, voucher: "Voucher", previous_value: None) -> None:
+        if not self.active:
+            return previous_value
+        self._trigger_voucher_event(WebhookEventAsyncType.VOUCHER_DELETED, voucher)
+
     def event_delivery_retry(self, delivery: "EventDelivery", previous_value: Any):
         if not self.active:
             return previous_value
         delivery_update(delivery, status=EventDeliveryStatus.PENDING)
         send_webhook_request_async.delay(delivery.pk)
+
+    def transaction_action_request(
+        self, transaction_data: "TransactionActionData", previous_value: None
+    ) -> None:
+        if not self.active:
+            return previous_value
+        event_type = WebhookEventAsyncType.TRANSACTION_ACTION_REQUEST
+        if webhooks := get_webhooks_for_event(event_type):
+            payload = generate_transaction_action_request_payload(
+                transaction_data, self.requestor
+            )
+            trigger_webhooks_async(
+                payload,
+                event_type,
+                webhooks,
+                subscribable_object=transaction_data,
+                requestor=self.requestor,
+            )
 
     def __run_payment_webhook(
         self,
@@ -950,6 +1000,11 @@ class WebhookPlugin(BasePlugin):
         )
 
     def is_event_active(self, event: str, channel=Optional[str]):
-        map_event = {"invoice_request": WebhookEventAsyncType.INVOICE_REQUESTED}
+        map_event = {
+            "invoice_request": WebhookEventAsyncType.INVOICE_REQUESTED,
+            "transaction_action_request": (
+                WebhookEventAsyncType.TRANSACTION_ACTION_REQUEST
+            ),
+        }
         webhooks = get_webhooks_for_event(event_type=map_event[event])
         return any(webhooks)